[MNG-6638] - Prevent reparsing POMs in MavenMetadataSource
If a dependency on another project is found, don't parse its
POM again, but instead use the model we already have in memory.

This saves a large amount of time and memory for builds that have
lots of subprojects and dependencies between them.

The approach was directly copied from DefaultArtifactDescriptorReader,
which does the same optimization.